[CL-2219] Move report builder to core
