+++ front/app/containers/Admin/reporting/components/ReportBuilder/Editor/index.tsx
