--- front/app/modules/commercial/report_builder/index.tsx
@@ -1,17 +1,0 @@
-import React from 'react';
-import { ModuleConfiguration } from 'utils/moduleUtils';
-
-const ReportBuilder = React.lazy(() => import('./admin/containers/index'));
-
-const configuration: ModuleConfiguration = {
-  routes: {
-    admin: [
-      {
-        path: 'report-builder',
-        element: <ReportBuilder />,
-      },
-    ],
-  },
-};
-
-export default configuration;
