+++ front/app/modules/index.ts
@@ -43,7 +43,6 @@ import eventsWidgetConfiguration from './commercial/events_widget';
 import insightsConfiguration from './commercial/insights';
 import analyticsConfiguration from './commercial/analytics';
 import representativenessConfiguration from './commercial/representativeness';
-import reportBuilderConfiguration from './commercial/report_builder';
 
 import idViennaSamlConfiguration from './commercial/id_vienna_saml';
 
@@ -206,9 +205,5 @@ export default loadModules([
   {
     configuration: impactTrackingConfiguration,
     isEnabled: CL_CONFIG['modules']['commercial/impact_tracking'],
-  },
-  {
-    configuration: reportBuilderConfiguration,
-    isEnabled: CL_CONFIG['modules']['commercial/report_builder'],
   },
 ]);
